Appeasing the React validator. Passing `null` for unused values instead of `false`.

+++ src/style_guide/components/data_table/template_td.js
@@ -53,7 +53,7 @@ class DataTableCell extends React.Component {
         <FormattedDate
           value={value}
           day='numeric'
-          month='long'
+          month='numeric'
           year='numeric'
         />
       )
